+++ resources/views/default/_partials/navigation/page.blade.php
@@ -1,6 +1,13 @@
+@php
+  $tooltip = '';
+  if (strlen (strip_tags($title)) > 15) {
+    $tooltip = 'title="' . strip_tags($title) . '"';
+  }
+@endphp
+
 @if($hasChild)
   <li class="nav-item has-treeview {!! ($isActive) ? 'menu-open' : '' !!}">
-    <a href="#" class="nav-link" {!! $attributes !!}>
+    <a href="#" class="nav-link" {!! $attributes !!} {!! $tooltip !!}>
       {!! $icon !!}
       <p class="{{ $icon ? 'ml-2':'' }}">
         {!! $title !!}
@@ -25,7 +32,7 @@
   </li>
 @else
   <li class="nav-item">
-    <a href="{{ $url }}" class="nav-link {!! ($isActive) ? 'active' : '' !!}" {!! $attributes !!}>
+    <a href="{{ $url }}" class="nav-link {!! ($isActive) ? 'active' : '' !!}" {!! $attributes !!} {!! $tooltip !!}>
       {!! $icon !!}
       <p class="{{ $icon ? 'ml-2':'' }}">
         {!! $title !!}